fix image-thumbnail functionality
Although the "image-thumbnail" usage statement was supported in
earlier versions of this script, the correct URLs would not have been
loaded into ArchivesSpace. This is because the publication
infrastructure assumes that image-thumbnail usage statement is bound
to a URL that includes a urlappend query.  e.g.,
http://hdl.handle.net/2333.1/foo?urlappend=/mode/thumb

This version of the script automatically appends the proper urlappend
string to the handle if the usage statement is 'image-thumbnail'.

ticket:
https://jira.nyu.edu/jira/browse/DLTSASPACE-19